pyproject.toml
[project]
name = "jute-disease-detection"
version = "0.1.0"
description = "Deep Learning Exploration for Jute Leaf Disease Detection"
readme = "README.md"
requires-python = "==3.11.14"
dependencies = [
    "albumentationsx[headless]>=2.0.14",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "ipykernel>=7.1.0",
    "kagglehub>=0.4.2",
    "lightning[pytorch-extra]>=2.6.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.0",
    "pip>=25.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "timm>=1.0.24",
    "torch",
    "torchinfo>=1.8.0",
    "torchvision",
    "wandb>=0.24.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jute_disease"]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "T20"]
ignore = ["N812"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-m 'not slow'"
markers = [
    "slow: marks tests as slow (deselected by default; run with -m slow)",
]
